small models changes
