+++ app/Models/Subjects/Organization.php
@@ -1,10 +1,11 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Models\Subjects\Organizations;
+namespace App\Models\Subjects;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Support\Carbon;
 
 /**
@@ -13,6 +14,9 @@ use Illuminate\Support\Carbon;
  * @property string $description
  * @property string $registration_number
  * @property Carbon $founding_date
+ * @property Clinic[] $clinics
+ * @property Doctor[] $doctorsStudied
+ * @property Pharmacy[] $pharmacies
  */
 class Organization extends Model
 {
@@ -28,6 +32,19 @@ class Organization extends Model
     protected $casts = [
         'founding_date' => 'datetime',
     ];
+
+    function clinics(): HasMany
+    {
+        return $this->hasMany(Clinic::class);
+    }
+    function doctorsStudied(): HasMany
+    {
+        return $this->hasMany(Doctor::class);
+    }
+    function pharmacies(): HasMany
+    {
+        return $this->hasMany(Pharmacy::class);
+    }
     public $timestamps = true;
 
 }
